dt-bindings: pwm: Add binding for Allwinner H616 PWM controller
The Allwinner H616 PWM controller is very similar to the D1 PWM
controller with slight register differences. Add a device tree
binding it. Note that while the H616 and H700 SoCs are identical,
there are no pins to output PWM channel 0 or 5 on the H616. In this
instance though the hardware still behaves as if it has 6 channels.

Signed-off-by: Chris Morgan <macromorgan@hotmail.com>